requirements.txt
python-dotenv
requests
networkx
pyahocorasick
google-generativeai
pypdf
PyQt6
cryptography
